+++ mod.ts
@@ -42,7 +42,7 @@ export default function lume(
   const site = new Site(options as Partial<SiteOptions>);
 
   // Ignore the .git folder and .DS_Store (macOS) files by the watcher
-  site.options.watcher.ignore.push(".git");
+  site.options.watcher.ignore.push("/.git");
   site.options.watcher.ignore.push((path) => path.endsWith("/.DS_Store"));
 
   return site